+++ lib/crypto.js
@@ -20,6 +20,7 @@ const constants = require('constants');
 const stream = require('stream');
 const util = require('util');
 const internalUtil = require('internal/util');
+const LazyTransform = require('internal/streams/LazyTransform');
 
 const DH_GENERATOR = 2;
 
@@ -40,37 +41,6 @@ exports._toBuf = toBuf;
 
 const assert = require('assert');
 const StringDecoder = require('string_decoder').StringDecoder;
-
-
-function LazyTransform(options) {
-  this._options = options;
-}
-util.inherits(LazyTransform, stream.Transform);
-
-[
-  '_readableState',
-  '_writableState',
-  '_transformState'
-].forEach(function(prop, i, props) {
-  Object.defineProperty(LazyTransform.prototype, prop, {
-    get: function() {
-      stream.Transform.call(this, this._options);
-      this._writableState.decodeStrings = false;
-      this._writableState.defaultEncoding = 'binary';
-      return this[prop];
-    },
-    set: function(val) {
-      Object.defineProperty(this, prop, {
-        value: val,
-        enumerable: true,
-        configurable: true,
-        writable: true
-      });
-    },
-    configurable: true,
-    enumerable: true
-  });
-});
 
 
 exports.createHash = exports.Hash = Hash;
